Example of co-log integrated into a transformer
This was done by storing the logger action in a ReaderT environment on
our end and using instancing into HasLog (not something seemingly more
conventional like a LoggerT).

Most of this code comes from the co-log github docs at
<https://github.com/co-log/co-log/blob/main/tutorials/2-custom/Custom.md>

So, it works but I found it confusing to write and don't love it as
compared with plainer logging solutions.

Closes gh-2